Simplified code and deleted unwanted comments
